+++ src/pywmd/__init__.py
@@ -1,0 +1,5 @@
+'''The md package - Read and write Markdown formatted documents
+
+This package contains modules
+
+'''