+++ src/sample/cnv_md.py
@@ -1,0 +1,32 @@
+"""Markdown converter for yWriter projects.
+
+This is a PyWriter sample application.
+
+Copyright (c) 2021 Peter Triesberger
+For further information see https://github.com/peter88213/PyWriter
+Published under the MIT License (https://opensource.org/licenses/mit-license.php)
+"""
+SUFFIX = ''
+
+import sys
+import os
+
+from pywriter.ui.ui_cmd import UiCmd
+from pywriter.ui.ui_tk import UiTk
+from pywriter.converter.yw_cnv_ui import YwCnvUi
+
+from pywmd.md_file_factory import MdFileFactory
+
+
+def run(sourcePath, suffix=None):
+    ui = UiTk('yWriter import/export')
+    converter = YwCnvUi()
+    converter.ui = ui
+    converter.fileFactory = MdFileFactory(
+        markdownMode=False, noSceneTitles=False)
+    converter.run(sourcePath, suffix)
+    ui.start()
+
+
+if __name__ == '__main__':
+    run(sys.argv[1], SUFFIX)